[RENAME] : 파일 위치 변경 main, login
