+++ app/src/main/java/com/example/frompet/login/LoginActivity.kt
@@ -1,7 +1,8 @@
-package com.example.frompet
+package com.example.frompet.login
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.frompet.R
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
